+++ hypothesaes/sae.py
@@ -411,9 +411,9 @@ def get_multiple_sae_activations(sae_list, X, return_neuron_source_info=False, *
 
     activations_list = []
     neuron_source_sae_info = []
-    for s in sae_list:
+    for (idx, s) in enumerate(sae_list):
         activations_list.append(s.get_activations(X, **kwargs))
-        neuron_source_sae_info += [(s.m_total_neurons, s.k_active_neurons)] * s.m_total_neurons
+        neuron_source_sae_info += [(f'SAE_{idx}', s.m_total_neurons, s.k_active_neurons)] * s.m_total_neurons
     activations = np.concatenate(activations_list, axis=1)
     
     if return_neuron_source_info: